quorum-tiebreaker: adapt to improvements in DRBD
In test case 2 A keeps quorum now (sometimes).

When DRBD manages to perform a two-phase-commit after A-B got blocked
and before B-C gets blocked.

On my home machine, it usually can not complete the 2PC and the test
works with the more recent DRBD as it is.

On C64, it usually can complete the 2PC and the test breaks since A
can not re-gain quorum after the A connects with B again.