ci: make vmshed quiet by default
This whole vmshed mechanism has been working pretty well for us for a
while now, so I don't think the debug outputs are generally useful
anymore.
On the other hand, we are running lots of pipelines and the debug output
is not only cluttering the logs, but the size of the logs is also adding
up.

So, supply "--quiet" to vmshed by default to hide the debug output.
If we need it again for some reason, it can be enabled for a pipeline by
setting VMSHED_ARGS to include "--quiet=false".