twopc-conflict-one-initiator: new test
This test examines the retry behavior when a TWO-PC needs to be
retried. It starts with:

     B
      \
   A   C

Then it establishes A-B, but makes sure that the P_TWOPC_COMMIT gets
delayed when it travels from B to C.
A then quickly establishes the connection to C. The goal of the test
is to provoke a conflict on C. C should still execute the previous
state change while it already gets the P_TWOPC_PREPARE for the next
state change.