drbdtest: terminate remote dmesg process
Otherwise it is left running on the remote machine. Terminating the SSH
client process is not sufficient to kill it.

This is necessary with lbpytest since commit:
controlmaster: read until stdout and stderr reach end-of-file

The "ssh" command has strange handling of file descriptors. It passes
stdout and stderr to the control master. When the non-master process is
terminated, the master process closes stdout, but not stderr. As a
result, SSH.pipeIO() does not return.