diskless-logical-block-size: Also test the 'block-size' config option
DRBD needs to combine the physical block sizes of the backing
disks. (A diskless node does not contribute a physical block size on
its own.) In clusters with storage devices with different physical
block sizes, it is necessary to configure the maximal physical block
sizes on the DRBD level. Here is an example highlighting the need.

Let's say node A is diskless. It connects to node B, which has a
physical block size of 512 bytes. Then the user mounts the filesystem
on node A; the filesystem recognizes that it can do I/O in units of
512 bytes.
Later, node C joins the cluster with a physical block size of 4096
bytes. Now, suddenly DRBD starts to deliver I/O errors to the
filesystem if it chooses to do I/O on e.g., 512 or 1024 bytes.

The new configure option "block-size" comes to the rescue. You can
configure DRBD on node A to advertise a hw_sector_size of 4096 to the
filesystem before such a node becomes visible.