upgrade: make test reliable with DRBD 8.4 and 9
When upgrading from DRBD 8.4 to 9 we need to re-write the DRBD config.
Otherwise the wrong node-id for the peer may be used, leading to test
failures:
Pattern 'peer-device .* replication:SyncTarget' ('volume:0', 'peer-node-id:0') does not match in lbtest-vm-43.test:2-29"

The above event is found in the logs for this run with peer-node-id:1.