virter: explicitly reinstall modules packages in base image provisioning
Prior to this change, when the source image from AlmaLinux has the same
kernel version as the one being installed, only the kernel package was
reinstalled. The modules packages were not reinstalled. The modules from
AlmaLinux are not entirely compatible with the kernel from RHEL with the
same version. This caused the VM to fail to boot.