resync-reconciliation: Improve test stability
This test was written long before drbd commit
9e4e702b5 drbd: initiate an empty two-phase-commit after a secondary peer crashed
Since that commit, the test contains a race condition between cutting
the connection to the second secondary and that two-phase-commit
triggered on the primary.

With unlucky timing, the two-pc gets executed by the two secondaries
when the connection to the initiator of the two-pc gets cut. The two
secondaries then need to wait until this two-pc times out.

By that time, the test terminates with an error.

A change that I am preparing for DRBD makes this unlucky timing to a
100% hit.

An alternative fix is adding 'twopc-timeout 50;' (5 seconds) to the
resource options.