new test: add-disk
The test adds new volumes to a resource that has already established
connections.

Test 1 first adds the new minor to all nodes and waits until every
node recognizes that the other nodes also added that minor. After
that, it continues with attaching the backing disks. This method
works with older DRBD releases.
(Probably, it was only tested this way)

Test 2 uses `adjust` on the nodes in parallel. This test fails up to
drbd-9.1.15 and drbd-9.2.4.

Test 3 skips the initial resync