drbdtest: ignore error when getting installed drbd package
When running the mainline tests, drbd will already be compiled into the
kernel. This means that there is no module to load and also no package
to install. So when we try to query the installed package, it will fail.

Currently, we don't use the drbd_versions.json file in the mainline test
suite at all, so it's fine for now to just store an empty string here.