new test: connect-differing-sizes
test 1:

Consider a cluster of 3 nodes with all possible connections
configured. The only obstacle is that one node (A) has a slightly
smaller backing disk, and it is primary before establishing the
connections.

       B
      / \
     A---C

The test connects A and B first, and then B and C. It verifies that
DRBD does not drop the established connections when the smaller size
propagates.

test 2:

In a linear chain of nodes

   A - B - C

The test verifies that the smaller size propagates through to node C
when A joins last.

Both test cases fail up to (including) drbd-9.1.17.