virter: exit with special code if drbd fails to load
When mass-testing drbd with lots of different kernels, we want to make
some qualitative assessment of how well the drbd module works with a
given kernel. For that, we need to know *why* the test suite failed.

As a first rough distinction, we can group test suite failures into
three categories:
1. Runtime failure
2. DRBD Load-time failure
3. Other system failure

For case 1, vmshed will exit successfully, but some or all tests will
fail.
Case 3 generally signalizes that the test suite itself is broken, and
vmshed will exit with a non-zero exit code.
Case 2 is what is covered here: no tests can be run because the DRBD
module is incompatible, but the test suite itself is fine. Signalize
this condition by exiting with a "magic" exit code, 100.