drbdtests: fix wait for resync progress loops
resync-target-promote had some weird behavior where it would sometimes not
terminate, claiming a log line was not found which clearly is present.

apparently that was because we told logscan to wait for
`peer_device .* done:(<percentage>)`,
for more than one peer device at a time,
but looping around when at least one such percentage was too small still.
If the other peer device was already completely done in the next iteration,
we would timeout waiting for its progress logline.

Fix that by NOT iterating, but changing the to be matched pattern.

While at it, change the same coding pattern in other tests as well, without
chaning their respective limits (sometimes 10, sometimes 20 percent done).
Even though that was not strictly necessary, as they seem to only scan for
a single peer device per iteration.