resync-w-dless-pri: Also test with I/O during the resync
Adding three new test cases.
This highlights that up to drbd-9.2.7:

In a cluster of A, B, C with A diskless and primary where the
connection from A to C is missing or down

   B
  / \
 A   C

And a resync running from B to C. Then the resync from B to C failed
to terminate if A executed a write after C sent its final resync
request to B and before B answered the last resync requests.