new-test: outdate-not-all
This test exercises some events in a three-node cluster with B and C
having a backing disk and A being diskless and primary.

    A
   / \
  B---C

Then the connection A-C gets interrupted. With that, C becomes Outdated.
A continues writing data.

    A
   /
  B---C

Now, the connection A-B dies. B starts a resync to C.

    A

  B-->C

While the resync is ongoing, A connects to C

    A
     \
  B-->C

Ensuring this does _not_ Outdate B.

DRBD up to 9.1.19 and 9.2.8 had this bug.