autopromote: restore stability and effectiveness
Commit c68c9d57762d ("replace open coded calls to setsid and fio with
busywrite") caused the test autopromote to become less reliable. The
test relied on fragile timing. With the switch to use fio instead of dd,
the time starting the process became longer, so that the signal was
often sent too early.

In addition, fio opens and closes the device once to perform the
BLKGETSIZE64 ioctl before it opens the device to actually write. As a
result, DRBD must auto-promote and auto-demote twice. This could cause
an event which makes the test pass even when the bug is present.

Fix these issues by switching from network delay to packet blocking.
This is a more robust way of controlling the timing. There are still
some sleeps involved, but these are less sensitive than the sleep in the
previous solution.

Fixes: c68c9d57762d ("replace open coded calls to setsid and fio with busywrite")