quorum-lost-uuid: new test
We are starting with a cluster of 3. Then B disconnects gracefully

    A          A
   / \   ->     \
  B---C      B   C

A promotes to primary and gets a write, so it also creates a new
current-UUID which it writes to disk. But the connection to C breaks
in just that moment. So the network packet informing C about it does
not make it to C. Then B and C connect.

    A

  B---C

B gets promoted to primary (B & C have quorum now). Strictly speaking, we
maneuvered it into a split-brain just that A is I/O frozen, and has not
modified its data set.
To recover Node A, we terminate the writing process using `drbdsetup
--force secondary`. Upon restoring the network between A and C, a
split-brain situation is detected! (A generated a new current-UUID
when it was isolated). Then, the test adds the 'after-sb-0pri
discard-zero-changes;' config option to the resources, allowing them
to reconnect.

    A          A
     \   ->   / \
  B---C      B---C