resync-reconciliation: allow separated out-of-sync and SyncSource lines
Since a recent DRBD change, the test resync-reconciliation-3 sometimes
fails. The test waits for:

peer-device .* replication:SyncSource .* out-of-sync:(\\d+)

But "out-of-sync" changes first, and so is not printed on the "SyncSource" line:

2024-04-23T20:31:44.852872+00:00 change peer-device name:res peer-node-id:1 conn-name:lbtest-vm-71 volume:0 peer-disk:Inconsistent peer-client:no done:0.00 received:0 sent:384 out-of-sync:128 pending:0 unacked:0 dbdt1:0.00 eta:nan
2024-04-23T20:31:44.852927+00:00 change peer-device name:res peer-node-id:1 conn-name:lbtest-vm-71 volume:0 replication:SyncSource

This is legitimate behavior from DRBD. Match these parts as separate
events so that they may appear on separate lines.