drbdtest: always capture active drbd version in dmesg
If DRBD was not loaded, modprobe will load it,
and the module init messages document the exact version.

If the module was already loaded prior to the start of the test,
this information may be lost (already rotated out of dmesg).

No longer: cat /proc/drbd > /dev/kmsg

If I accidentally loaded the known broken module version,
this may help avoid time spent trying to unnecessarily debug
what is actually expected...