quorum: detect quorum loss event before or after connection loss event
Since DRBD commit 121167a75456 ("drbd: add "dev_is_open" to device_info,
and broadcast on change"), the test quorum-3 has been less stable. A
side effect of that commit is that the device event for quorum loss can
occur before the connection loss events. Update the test to cope with
that.