interrupt-detach: new test
If a backing device suddenly ceases delivering I/O completions, and
in reaction, the user issues a `drbdsetup detach`, the operation
will hang when it tries to write internal meta-data.

Verify that the hanging detach operation can be interrupted with
a signal (Ctrl-C).
This empowers the user to continue with `drbdseup --force detach`.