outdate-early: use file system
Expose a race between creation of new uuid and replication link loss
causing split brain.

Until now, the test did a new open() of the DRBD,
which may delay the new write considerably.

13ada1be2 (drbd: open: do not delay open() if already Primary, 2024-10-02)
caused the timing to be quicker, exposing the problem more often.

Use a Filesystem, which keeps drbd open.
The expectation is that this will
expose the problem even more often.