new-test: twopc-abort-connect2
This test provokes the concurrent two-phase commit for establishing
the connections from node_a to node_c and node_a to node_b. At first,
the two-pc a-c starts. Then node_c aborts it and allows a-b to
overtake.

Up to drbd-9.1.22 and 9.2.11, this results in node_c having a resync
started and node_a having the connection established (without resync).