outdate: remove wait_backward_propagating_resyncs()
A Python `range(start, stop)` with `start >= stop` is an empty sequence.
Hence wait_backward_propagating_resyncs() did nothing.  To make it do
what was intended, we would need to set `step=-1`. However that makes
the test fail because we wait for other events on last_node which occur
concurrently. Instead of adding complicated event handling to match the
concurrent events, just remove the ineffective code.