outdate: move lost_primary_after_no_writes() to its own test
outdate_test() does not wait for the last node to become UpToDate. So
when `outdate_test(write_between_actions=True)` returns, a resync may
still be running. This can cause lost_primary_after_no_writes() to fail.
That function waits for the last node to become Outdated, which doesn't
happen if it was not UpToDate in between.

Unfortunately, waiting for UpToDate on last_node reveals an obscure DRBD
bug. Testing this is not a priority at the moment, so just work around
the issue by separating outdate into 2 tests.