outdate-not-all, quorum-isolate-primary, resync-reconciliation: stability
These tests block (at least) one connection from a primary and expect
to be able to stuff data into the socket. So far, these tests did that
directly after promoting the primary.
Most of the time, a freshly promoted primary needs to create a new
current-UUID before writing and stuffing data into a socket.
Add an explicit write after promoting the node to primary before
blocking one connection, which creates the new current-UUID.
With that, the intended stuffing of data into the socket works
reliably.