resync-reconciliation: improve test stability
With the recent changes in DRBD, starting a reconciliation resync
became quicker. With that, it became more likely that the thread
handling the connection to the primary would not make it to the
connection-state Unconnected before the reconciliation resync started.

With the event('connection:Unconnected', ...) in place, the event
machinery skips over the start of resync we look for in the next step.

Waiting for the NetworkFailure connection-state is sufficient.