resync-reconciliation: reduce data amount
Sometimes, the initial socket send buffer size was not large enough to
hold 128KiB. It was a race condition between the kernel increasing the
socket send buffer size and the primary node recognizing that the
"slow secondary" connection was blocked. Then, the resync is an "after
unstable" resync.

By reducing the amount to 64KiB, the test reliably tests the
reconciliation resync.