resync-reconciliation: fix wait condition
When waiting for the peers to become UpToDate, the condition
implicitly assumed that the two secondaries finished their resync
before and were already UpToDate when connecting to the (former)
primary.

When one of them is consistent during connect and changes its disk
state to UpToDate a moment later, the event looks like this:

change peer-device ....volume:0 peer-disk:UpToDate....

if it is already UpToDate during connect it looks like this:

change peer-device  ....volume:0 replication:Established peer-disk:UpToDate...

Fix the regex to deal with both cases.