autopromote: improve following DRBD state changes
Improve following the DRBD states so that in case the test fails, it
is easier to understand where the test failed.
Follow the first aspect ("Test interrupting a process while it
auto-promotes a drbd device") with the forbidden patterns active.

Increase the ping timeout for the first aspect to run reliably in the
test suite.