resync-from-weak: improve stability
When I execute the test locally, the resync is slow compared to the
test's progress. The timing sometimes differs when it runs in the
stability test loop, causing a percentage of the runs to fail. The
failing runs demonstrate a genuine bug in DRBD. I have created
drbd#474 to track that bug.

Fix the test by increasing the amount to rsync so that the test is
stable in the aspect it should test now.