resync-w-dless-pri: increase resync rate
Sometimes DRBD fails to complete the resync after connecting to the
Primary in the "test3(ResyncType.ONGOING_WRITES)" case. It ends up
running at a rate of dbdt1:0.33 for whatever reason. Since the backing
device size is rounded up to 12MiB, this causes the resync to take
longer than 30s. Increase the resync rate to eliminate these irrelevant
failures.