disktools.py: Consider modern day LVM
Recent LVM introduces the devicesfile, which deprecates the previous
filter options. With the new version, it is necessary to set the
devices/search_for_devnames setting to "none" to avoid LVM scanning
(a potentially IO-suspended) DRBD device.
The modern LVM ignores the filter and global_filter, and warns if they
are set.

I noticed this issue while using the test resync-snapshot on Fedora 42
target systems.