new test: resync-var-bm-bsize
This test creates two DRBD peers with one having a bitmap block size
of 8k and the other one 64KiB. After disconnecting the peers, it
writes a single 4k block every 64k with all possible offsets in that 64k
range, and lets it resync.
It also tests the basic behavioural properties of the --bitmap=no setting.