new test: quorum-suspend-failover-dless
This test creates a fully connected three-node cluster in which A
and B have backing disks, and C is a diskless node.

 A---B
  \ /
   C

Initially, node B is primary. After it gets isolated from A and B,
it freezes I/O. Then the diskless node C becomes a new primary node
and creates a new current UUID. Then A and B reestablish the
connection. B realizes it is outdated and forces itself into
secondary role.

The last step was not implemented in DRBD up to (including) 9.2.15.